fix: #108 log int -> long 변경

+++ src/main/java/com/purpleprint/network/purpleprintproject/analysis/command/application/dto/LogDTO.java
@@ -29,5 +29,5 @@ public class LogDTO {
     private Float xCoord;
     @JsonProperty("zCoord")
     private Float zCoord;
-    private int currentTime;
+    private long currentTime;
 }
